Add testing
This commit adds support for Rubocop and makes changes based on its
feedback. It also adds rspec tests that test the defaults for each
supported platform and does some type validation, though that is
incomplete. Further tests are needed for type validation as well as
logic and file content checks.